oups

+++ lib/glory/achievements.rb
@@ -20,11 +20,11 @@ module Achievements
     achievements.first(:conditions => conditions).present?
   end
   
-  def delete_achievement?(achievement, level = nil)
+  def delete_achievement(achievement, level = nil)
     conditions = {:type => achievement.to_s, :achievable_id => id, :achievable_type => self.class.to_s}    
     conditions[:level] = level if level
     lost_achievement = achievements.first(:conditions => conditions)
-    lost_achievement.destroy
+    lost_achievement.destroy if lost_achievement
   end
   
   def get_badges_in_progress(badges)
